% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19SMHvalidation.R
\docType{package}
\name{covid19SMHvalidation}
\alias{covid19SMHvalidation}
\title{covid19SMHvalidation: COVID-19 Scenario Modeling Hub Validations & Visualizations}
\description{
The COVID-19 Scenario Modeling Hub Validations package contains functions for validation and visualization scenario modeling hub submissions.
}
