% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covidcast_data.R
\name{fix_negative_counts}
\alias{fix_negative_counts}
\title{Add missing dates, fix counts that go negative, and fix NA values}
\usage{
fix_negative_counts(
  df,
  cum_col_name,
  incid_col_name,
  date_col_name = "Update",
  min_date = min(df[[date_col_name]]),
  max_date = max(df[[date_col_name]]),
  type = "mid"
)
}
\arguments{
\item{df}{data frame}

\item{cum_col_name}{character vector}

\item{incid_col_name}{character vector}

\item{date_col_name}{character vector, by default "Update"}

\item{min_date}{Date, by default `min(df[[date_col_name]])`}

\item{max_date}{Date, by default `max(df[[date_col_name]])`}

\item{type}{character vector: must be one of ['low', 'med', 'high']. By
default `"mid"`

#' @details There are three different ways of dealing with negative incidence
counts, specified by the type argument:
\itemize{
 \item{"high" method: }{Highest estimate; this modifies the cumulative column
  to be the cummax of the previous cumulative counts.}
 \item{"low" method: }{Assume all negative incidents are corrections to
 previous reports. So, reduce previous reports to make incidents not
 decrease.}
 \item{"mid" method: }{Combination of both low and high.}
}
For consecutive negatives, the first negative will be modified so that the
  cumulative is the cummax so far.
For the following negatives, use the low method: reduce previous reports.
For get_USAFacts_data(), this is always the default ("mid"), but the other
 code is here for posterity.}
}
\value{
a data frame
}
\description{
See fix_negative_counts_single_geoid() or the details section for more
details on the algorithm, specified by argument "type"
}
