% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{format_gs_data}
\alias{format_gs_data}
\title{Format Gold Standard Data for Plotting}
\usage{
format_gs_data(lst_gs, projection_date)
}
\arguments{
\item{projection_date}{}
}
\value{

}
\description{
Format Gold Standard Data for Plotting
}
