% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covidcast_data.R
\name{get_covidcast_data}
\alias{get_covidcast_data}
\title{Get "covidcast" data}
\usage{
get_covidcast_data(
  geo_level = "state",
  signals = c("deaths_incidence_num", "deaths_cumulative_num",
    "confirmed_incidence_num", "confirmed_cumulative_num",
    "confirmed_admissions_covid_1d"),
  limit_date = Sys.Date(),
  fix_negatives = TRUE,
  run_parallel = FALSE,
  n_cores = 4
)
}
\arguments{
\item{geo_level}{character, geographical level ("state" or "county").
By default, `"state"`}

\item{signals}{character, one or multiple signals accepted by the covidcast
package. By defaults: `c("deaths_incidence_num", "deaths_cumulative_num",
 "confirmed_incidence_num", "confirmed_cumulative_num",
 "confirmed_admissions_covid_1d")`}

\item{limit_date}{Date, by default, `Sys.Date()`}

\item{fix_negatives}{Boolean, by default TRUE. For more information, please
look at the documentation of the function `?fix_negative_counts()`}

\item{run_parallel}{Boolean, by default FALSE. If TRUE, will run the call
to the covidcast package in parallel.}

\item{n_cores}{numeric, by default 4. Number of cores to use when
run_parallel is set to TRUE}
}
\value{
a data frame
}
\description{
Download and returns data from COVIDcast package, with daily data available
at multiple geographical level.
}
\details{
For `covidcast` R package. Please, look at
\url{https://cran.r-project.org/web/packages/covidcast/index.html} for more
information. \cr\cr
}
\examples{
df <- get_covidcast_data()

}
