% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{make_state_plot_pdf}
\alias{make_state_plot_pdf}
\title{Generate PDF of state plots}
\usage{
make_state_plot_pdf(
  proj_data,
  gs_data,
  team_model_name,
  projection_date,
  save_path,
  plot_quantiles = c(0.025, 0.975),
  y_sqrt = FALSE
)
}
\arguments{
\item{y_sqrt}{}
}
\value{

}
\description{
Generate PDF of state plots
}
\examples{

}
