% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_projections}
\alias{plot_projections}
\title{Plot Projections}
\usage{
plot_projections(data, st, projection_date, legend_rows = 1, y_sqrt = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{st}{vector}

\item{projection_date}{a data vector}

\item{legend_rows}{numeric, 1 by default}

\item{y_sqrt}{boolean, FALSE by default}
}
\description{
Plot Projections
}
