% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gs_data.R
\name{pull_gs_data}
\alias{pull_gs_data}
\title{Pull Truth data from COVIDCast}
\usage{
pull_gs_data()
}
\value{
a list of dataframe, each dataframe named with the corresponding
covidcast signal except "hospitalization" instead of "confirmed_admissions_covid_1d"
}
\description{
Download and returns data from COVIDcast package in a Scenario Modeling
Hub format, with the incidence and cumulative data for each state and
US national level by epiweek.
Source data:
}
\examples{
pull_gs_data
}
