% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scen_round_info.R
\name{scen_round_info}
\alias{scen_round_info}
\title{Create summary table with round, scenario and date information}
\usage{
scen_round_info(repo_path = "midas-network/covid19-scenario-modeling-hub/")
}
\arguments{
\item{repo_path}{character vector, relative path of the COVID-19 scenario
model hub (i.e "midas-network/covid19-scenario-modeling-hub/")}
}
\description{
From a path linking to the Scenario Modeling Hub, reads all the scenario's
README and extracts scenario name, scenario id and 1 weak ahead projection
date information.
}
\examples{
scen_round_info()

}
