% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMHvalidation.R
\docType{package}
\name{SMHvalidation}
\alias{SMHvalidation-package}
\alias{SMHvalidation}
\title{SMHvalidation: Scenario Modeling Hub Validations & Visualizations}
\description{
The Scenario Modeling Hub Validations package contains functions for
validation and visualization scenario modeling hub submissions.
}
\author{
\strong{Maintainer}: Lucie Contamin \email{contamin@pitt.edu} (\href{https://orcid.org/0000-0001-5797-1279}{ORCID})

Authors:
\itemize{
  \item Shaun Truelove
  \item Scenario Modeling Hub Coordination Team
}

}
\keyword{internal}
