% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_agegroup.R
\name{test_agegroup}
\alias{test_agegroup}
\title{Runs Validation Checks on the \code{age_group} column}
\usage{
test_agegroup(df, model_task, file_name = "")
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.) for a specific round}

\item{file_name}{optional, name of the file tested}
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Validate Scenario Modeling Hub submissions: test if the \code{age_group} column
contains the expected  value.
}
\details{
This function  tests:
\itemize{
\item Age group: If the submission contains projection by age group,
the \code{age_group} column contains the age group written \verb{<AGEMIN>-<AGEMAX>}.
\item Age group value: If the submission contains projection by
age group, the \code{age_group} column contains the age group values as
specify in the associated SMH GitHub Repository.
\item Age group target: If the submission contains projection by
age group and if one or multiple targets required specific \code{age_group}
value(s), no additional value(s) is provided in the submission file.
}

\code{model_task} should match a specific round model tasks from the
\code{tasks.json} associated with the hub. The json is expected to follow
the \href{https://hubverse.io/en/latest/user-guide/hub-config.html}{hubverse}
schema at least version 5.0

As the function was deprecated, it will not be updated anymore. It was
updated a last time to match the 5.0 hubverse schema version. However, it
might returns duplicated message output.
}
