% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_location.R
\name{test_location}
\alias{test_location}
\title{\strong{Deprecated} - Runs Validation Checks on the Location column}
\usage{
test_location(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
Validate Scenario Modeling Hub submissions: test if the  \code{location} column
contains the expected  value.
}
\details{
This function contains a test:
\itemize{
\item Specific location: For the target(s) requiring only specific
location(s), no additional location is provided in the submission file
}

Function called in the \code{validate_submission()} function.
}
