% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_modelprojdate.R
\name{test_modelprojdate}
\alias{test_modelprojdate}
\title{Runs Validation Checks on the Model Projection Date column}
\usage{
test_modelprojdate(df, path, start_date)
}
\arguments{
\item{df}{data frame to test}

\item{path}{character vector path of the file being tested}

\item{start_date}{corresponds to the "1 wk ahead" target in the projection
file}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`model_projection_date` column contains one unique date value corresponding
to the projection start date of the round and also corresponds to the date
in the name of the submission file.
}
\details{
This function contains 4 tests:
\itemize{
 \item{Date value: }{The `model_projection_date` column contains date value}
 \item{Unique value: }{The `model_projection_date` column contains a unique
 value.}
 \item{Correspondance to the name of the file: }{The `model_projection_date`
 column contains a unique date value matching the date in the name of the
 submission file}
  \item{Correspondance to the projection starting date: }{The
 `model_projection_date` column contains a unique date value matching the
 projection starting date of the corresponding round}
}
Function called in the `validate_submission()` function.
}
