% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_origindate.R
\name{test_origindate}
\alias{test_origindate}
\title{Runs Validation Checks on the Origin Date column}
\usage{
test_origindate(df, path, id)
}
\arguments{
\item{df}{data frame to test}

\item{path}{character vector path of the file being tested}

\item{id}{character date "id" of the corresponding round}
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Validate Scenario Modeling Hub submissions: test if the
\code{origin} column contains one unique date value corresponding
to the expected value, in the expected format.
}
\details{
This function tests:
\itemize{
\item Unique value: The \code{origin_date} column contains an unique expected
value.
\item Format: The \code{origin_date} column contains a unique date value in either a
character ("YYYY-MM-DD") or a Date format. If the date is in a datetime
format, a warning will be returned
}

As the function was deprecated, it will not be updated anymore. It was
updated a last time to match the 5.0 hubverse schema version.
}
