% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_quantile.R
\name{test_quantiles}
\alias{test_quantiles}
\title{Runs Validation Checks on the Quantiles type and value columns}
\usage{
test_quantiles(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Validate Scenario Modeling Hub submissions: test if the
\code{quantile} and \code{type} columns contain the expected information and value
and the projection value increases with the quantiles.
}
\details{
This function tests:
\itemize{
\item Type: The submission contains quantiles value when expected.
\item Quantiles: The \code{quantile} column matches the expected quantiles.
\item Required quantiles: The projection should contain all required
quantiles. It is accepted to submit less quantiles but the
function will return a warning and the submission will not be included
in the Ensembles.
\item Value: The \code{value} associated with each \code{quantile} is increasing as
the quantile increased. For example, if quantile 0.01 = 5 than quantile
0.5 should be equal or greater than 5.
\item Target: All targets required to have quantiles information have all
required quantiles for each group id (scenario/location/horizon/etc.).
}

\code{model_task} should match a specific round model tasks from the
\code{tasks.json} associated with the hub. The json is expected to follow
the \href{https://hubverse.io/en/latest/user-guide/hub-config.html}{hubverse}
schema at least version 5.0

As the function was \emph{deprecated}, it will not be updated anymore. It was
updated a last time to match the 5.0 hubverse schema version. However, it
might returns duplicated message output or false error as for example
the 5.0 version of the \code{tasks.json} does not accept both required and
optional quantile for the same task ids and past SMH rounds might contains
required and optional quantiles for optional and/or required target.
}
