% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_raceethnicity.R
\name{test_raceethnicity}
\alias{test_raceethnicity}
\title{Runs Validation Checks on the \code{race_ethnicity} column}
\usage{
test_raceethnicity(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Validate Scenario Modeling Hub submissions: test if the \code{race_ethnicity}
column contains the expected  value.
}
\details{
This function tests:
\itemize{
\item Race ethnicity value: If the submission contains projection by
race/ethnicity, the \code{race_ethnicity} column contains the expected values as
specified.
\item Race ethnicity tasks id: If the submission contains projection by
race/ethnicity and if one or multiple tasks ids required specific
\code{race_ethnicity} value(s), no additional value(s) is provided in the
submission file.
}

\code{model_task} should match a specific round model tasks from the
\code{tasks.json} associated with the hub. The json is expected to follow
the \href{https://hubverse.io/en/latest/user-guide/hub-config.html}{hubverse}
schema at least version 5.0

As the function was deprecated, it will not be updated anymore. It was
updated a last time to match the 5.0 hubverse schema version. However, it
might returns duplicated message output.
}
