% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_req_value.R
\name{test_req_value}
\alias{test_req_value}
\title{Runs Validation Checks on the required value}
\usage{
test_req_value(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)

Function called in the \code{validate_submission()} function, only if the
submission contains \code{"sample"} output type.}
}
\description{
Validate Scenario Modeling Hub submissions: test if all the required value
are present in the submission file
}
