% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sample.R
\name{test_sample}
\alias{test_sample}
\title{Runs Validation Checks on the sample column}
\usage{
test_sample(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
Validate Scenario Modeling Hub submissions: test if the  `sample` column
contains the expected  value.
}
\details{
This function contains 3 tests:
\itemize{
 \item{sample value: }{The submission should contain a sample column with
 value between 0 and 100.}
  \item{unique sample: }{The submission should contain a unique sample
  identifier for each scenario/target/location (age_group) group.}
   \item{sample integer: }{The submission should contain a sample column with
 integer only.}
}
Function called in the `validate_submission()` function.
}
