% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_submission.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate SMH (Scenario Modeling Hub) Submissions}
\usage{
validate_submission(
  path,
  js_def,
  hub_path,
  target_data = NULL,
  pop_path = NULL,
  merge_sample_col = NULL,
  partition = NULL,
  n_decimal = NULL,
  round_id = NULL,
  verbose = TRUE,
  r_schema = NULL
)
}
\arguments{
\item{path}{path to the submissions file (or folder for partitioned data)
to test, or string of parquet files (in this case, the validation will be
run on the aggregation of all the parquet files together, and each file
individually should match the expected SMH standard).
If partition is not set to \code{NULL}, path to the folder containing ONLY the
partitioned data.}

\item{js_def}{path to JSON file containing round definitions: names of
columns, target names, ... following the \code{tasks.json}
\href{https://hubdocs.readthedocs.io/en/latest/user-guide/hub-config.html}{Hubverse }
format}

\item{hub_path}{path to the repository contains the submissions files
and js_def file}

\item{target_data}{data frame containing the
observed data. We highly recommend to use the output of the
\code{pull_gs_data()} function. The data frame should have the time
series hubverse format.
Set to \code{NULL} (default), to NOT run comparison with observed data.}

\item{pop_path}{path to a table containing the population size of each
geographical entities by FIPS (in a column \code{"location"}).
Use to compare that value is not higher than expected population size.
Set to \code{NULL} (default), to NOT run comparison with population data.}

\item{merge_sample_col}{vector to indicate if the for the output type
"sample", the output_type_id column is set to NA and the information is
contained into other columns. If so the parameter should be set to the sample
ID columns names, for example: \verb{c("run_grouping" and "stochastic_run")}.
By default, \code{NULL} (sample ID information in the output type id column)}

\item{partition}{vector, for csv and parquet files, allow to validate files
in a partition format, see \code{arrow} package for more information, and
\code{arrow::write_dataset()}, \code{arrow::open_dataset()} functions.  By default
\code{NULL}, no partition}

\item{n_decimal}{integer,  number of decimal point accepted in the column
value (only for \code{"sample"} output type), if \code{NULL} (default) no limit
expected.}

\item{round_id}{character string, round identifier. If \code{NULL} (default),
extracted from \code{path}.}

\item{verbose}{Boolean, if TRUE add information about the sample pairing
information in output message. By default, \code{TRUE} (slows the validation
validation for sample output type)}

\item{r_schema}{Schema arrow objects, to read partition files with a specific
schema. If none provided (default, \code{NULL}), the schema will be created from
the \code{js_def} JSON file. (only for partition files)}
}
\description{
Runs all the different validation checks functions (\code{test_column()},
\code{test_scenario()}, \code{test_origindate()}, \code{test_type()}, \code{test_val()},
\code{test_target()}, \code{test_location()}, \code{test_other_col()}) on SMH submission
file(s) and prints information about the results of each tests on the
submission file(s): warning(s), error(s) or message if all the tests were
successful.
}
\details{
For more information on all tests run on the submission, please refer
to the documentation of each \verb{"test_*}" function. A vignette with all the
information is available in the package and is called:
vignette("validation-checks").

The function accepts submission in PARQUET, CSV, ZIP or GZ file formats. If
the submission files is in a "partitioned" format, the \code{path} parameter
should be to a directory to a folder containing ONLY the "partitioned"
files. If any other file is present in the directory, it will be included
in the validation.

The function runs some preliminary tests before calling the "test_*"
functions:
\itemize{
\item Input submission file format: The file format of the submission
file(s) correspond to the expected format (for example: \code{parquet}, or \code{csv},
etc.). If multiple files inputted, only \code{parquet} is accepted.
\item Date information: The column \code{origin_date} in the submission file
corresponds to a \code{model_tasks} round information in the JSON file
(\code{js_def} parameter).
\item Date format: All columns containing dates information should be in
"YYY-MM-DD" format.
}
}
\examples{
\dontrun{
validate_submission("PATH/TO/SUBMISSION", "PATH/TO/ROUND/tasks.json")
}
}
