% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{generate_validation_plots}
\alias{generate_validation_plots}
\title{Generate Validation Plots}
\usage{
generate_validation_plots(
  path_proj,
  lst_gs,
  save_path = dirname(path_proj),
  y_sqrt = FALSE,
  plot_quantiles = c(0.025, 0.975)
)
}
\arguments{
\item{lst_gs}{}
}
\value{

}
\description{
Generate Validation Plots
}
\examples{

}
