% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{make_state_plot_pdf}
\alias{make_state_plot_pdf}
\title{Generate PDF of state plots}
\usage{
make_state_plot_pdf(
  proj_data,
  gs_data,
  team_model_name,
  projection_date,
  save_path,
  plot_quantiles = c(0.025, 0.975),
  y_sqrt = FALSE
)
}
\arguments{
\item{proj_data}{dataframe}

\item{gs_data}{list of dataframe, containing observed data}

\item{team_model_name}{character vector, name of the team and models,
following Scenario Modeling Hub standard}

\item{projection_date}{date vector}

\item{save_path}{character vector, path to the saving folder for the PDF
output}

\item{plot_quantiles}{numeric vector, quantiles to use for plotting (should
correspond to the quantiles from the `proj_data` parameter)}

\item{y_sqrt}{boolean, FALSE by default}
}
\description{
Generate PDF of state plots
}
