% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{print_table}
\alias{print_table}
\title{Print outlier tables for validation report}
\usage{
print_table(
  data = proj_plot_data_calib_cum,
  tab_title = "Top Outliers: Percent difference from ground truth, Cumulative",
  metric = "prctdiff_gt",
  thresholds = c(-Inf, -0.2, -0.06, 0),
  colors = c("red", "orange", "yellow", "orange")
)
}
\arguments{
\item{data}{data frame}

\item{tab_title}{character vector, title}

\item{metric}{character vector, metric}

\item{thresholds}{numeric vector, thresholds}

\item{colors}{vector, colors}
}
\description{
Print outlier tables for validation report
}
