% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gs_data.R
\name{pull_gs_data}
\alias{pull_gs_data}
\title{Pull Truth data from COVIDCast}
\usage{
pull_gs_data(
  signals = c("deaths_cumulative_num", "confirmed_cumulative_num",
    "confirmed_incidence_num", "deaths_incidence_num", "confirmed_admissions_covid_1d")
)
}
\arguments{
\item{signals}{list of signals to download (should be one or multiple in
the default list)}
}
\value{
a list of dataframe, each dataframe named with the corresponding
covidcast signal except "hospitalization" instead of
"confirmed_admissions_covid_1d"
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Download and returns data from COVIDcast package in a Scenario Modeling
Hub format, with the incidence and cumulative data for each state and
US national level by epiweek.
}
\details{
The function output a list of 5 data frame (one by target) of observed
data in in a Scenario Modeling Hub format. The data are used on the
Scenario Modeling Hub website and are regenerated by using the \code{covidcast}
R package. Please, look at
\url{https://cran.r-project.org/web/packages/covidcast/index.html}
for more information.
\itemize{
\item For the cumulative numbers: use the \code{confirmed_cumulative_num} and
\code{deaths_cumulative_num} signals from \code{jhu-csse}. As the data are expressed
per day and the Scenario Modeling Hub use cumulative data by epiweek, the
last day of each epiweek is selected for each location and the values are
added between all states to obtains the US level data.
\item For the incidence numbers: use the \code{confirmed_incidence_num} and
\code{deaths_incidence_num} signals from \code{jhu-csse} and
\code{confirmed_admissions_covid_1d} from \code{hhs}. As the data are expressed per
day and the Scenario Modeling Hub use cumulative data by epiweek, the sum
of the incidences is calculated by epiweek and by location, and the values
are added between all states to obtains the US level data.
}

\strong{Source data:}

For case/death data, we use:
\itemize{
\item https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html # nolint
\item Signals: \code{"deaths_cumulative_num"}, \code{"confirmed_cumulative_num"},
\code{"confirmed_incidence_num"}, \code{"deaths_incidence_num"}
}

For hospitalization data, we use:
\itemize{
\item https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/hhs.html
\item Signals: \code{"confirmed_admissions_covid_1d"}
}
}
\examples{
\dontrun{
 lst_df <- pull_gs_data()
 lst_df
 str(lst_df)
}

}
