% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gs_data.R
\name{pull_gs_data}
\alias{pull_gs_data}
\title{Pull Truth data from COVIDCast}
\usage{
pull_gs_data()
}
\value{
a list of dataframe, each dataframe named with the corresponding
covidcast signal except "hospitalization" instead of
"confirmed_admissions_covid_1d"
}
\description{
Download and returns data from COVIDcast package in a Scenario Modeling
Hub format, with the incidence and cumulative data for each state and
US national level by epiweek.
}
\details{
The function output a list of 5 data frame (one by target) of observed
data in in a Scenario Modeling Hub format. The data are used on the
Scenario Modeling Hub website and are regenerated by using the `covidcast`
R package. Please, look at
\url{https://cran.r-project.org/web/packages/covidcast/index.html}
for more information.

* For the cumulative numbers: use the `confirmed_cumulative_num` and
  `deaths_cumulative_num` signals from `jhu-csse`. As the data are expressed
  per day and the Scenario Modeling Hub use cumulative data by epiweek, the
  last day of each epiweek is selected for each location and the values are
  added between all states to obtains the US level data.

* For the incidence numbers: use the `confirmed_incidence_num` and
  `deaths_incidence_num` signals from `jhu-csse` and
  `confirmed_admissions_covid_1d` from `hhs`. As the data are expressed per
  day and the Scenario Modeling Hub use cumulative data by epiweek, the sum of
  the incidences is calculated by epiweek and by location, and the values are
  added between all states to obtains the US level data.

**Source data:**

For case/death data, we use:
\itemize{
    \item{Link: }{\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/jhu-csse.html}}
    \item{Signals: }{"deaths_cumulative_num", "confirmed_cumulative_num",
    "confirmed_incidence_num", "deaths_incidence_num" }
}

For hospitalization data, we use:
\itemize{
    \item{Link: }{\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals/hhs.html}}
    \item{Signals: }{"confirmed_admissions_covid_1d"}
}
}
\examples{
lst_df <- pull_gs_data()
lst_df
str(lst_df)
}
