% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read CSV, ZIP, GZ pr PQT files}
\usage{
read_files(path, na_string = c("", "NA", "NaN"), sep = ",")
}
\arguments{
\item{path}{path of the file which the date are to be read from}

\item{na_string}{character vector to be interpreted as \code{NA} values}

\item{sep}{character, separator character between columns}
}
\description{
Reads a file and output a data frame. The function can read files in CSV,
ZIP, GZ and PQT format
}
\examples{
\dontrun{
 df <- read_files("PATH/TO/FILE")
}

}
