% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scen_round_info.R
\name{scen_round_info}
\alias{scen_round_info}
\title{Create summary table with round, scenario and date information [DEPRECATED]}
\usage{
scen_round_info(
  repo_path = "midas-network/covid19-scenario-modeling-hub/",
  branch = "master",
  add_readme = NULL
)
}
\arguments{
\item{repo_path}{character vector, relative path of the COVID-19 scenario
model hub (i.e "midas-network/covid19-scenario-modeling-hub/")}

\item{branch}{character, name of the branch to extract information from. By
default "master"}

\item{add_readme}{data frame of 4 columns: files (NA),
url ("https""//LINK/README") and round ("round1") to manually add README
path information in the output. By default, NULL (does not add anything)}
}
\description{
From a path linking to the Scenario Modeling Hub, reads all the scenario's
README and extracts scenario name, scenario id and 1 weak ahead projection
date information.
}
\details{
For the four first rounds of COVID-19, the SMH README are slightly
different and required the use of the "add_readme" parameters or it will
returns an error. Please see the examples

Function deprecated in Feb, 2023.
}
\examples{

\dontrun{
# For COVID-19
repo_path <- "https://raw.githubusercontent.com/midas-network/covid19-scenario-modeling-hub/"
scen_round_info(add_readme =  data.frame(
  files = NA,
  url = c(
    paste0(repo_path, "6dc683b9710c3a7eee6fa40d2986d0d79c0c918f/README.md"),
    paste0(repo_path, "a88188d964543ae06885d499eecaba8e34ee68f9/README.md"),
    paste0(repo_path, "52263c2086d8ef2d59c92b0a72d0d5c521290917/README.md")
  ),
  round = c("round1", "round2", "round3")
)
)

# For FLU
scen_round_info(repo_path = "midas-network/flu-scenario-modeling-hub/",
               branch = "main")
}

}
