% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports-hubValidations.R
\name{store_msg_val}
\alias{store_msg_val}
\title{Store validation output in a simple object}
\usage{
store_msg_val(msg)
}
\arguments{
\item{msg}{validation output}
}
\description{
Store validation output in a simple object
}
