% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_agegroup.R
\name{test_agegroup}
\alias{test_agegroup}
\title{Runs Validation Checks on the `age_group` column}
\usage{
test_agegroup(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
Validate Scenario Modeling Hub submissions: test if the `age_group` column
contains the expected  value.
}
\details{
This function contains 3 tests:
\itemize{
 \item{Age group: }{If the submission contains projection by age group,
 the `age_group` column contains the age group written `<AGEMIN>-<AGEMAX>`.}
 \item{Age group value: }{If the submission contains projection by
 age group, the `age_group` column contains the age group values as
 specify in the associated SMH GitHub Repository.}
 \item{Age group target: }{If the submission contains projection by
 age group and if one or multiple targets required specific `age_group`
 value(s), no additional value(s) is provided in the submission file.}
}
Function called in the `validate_submission()` function.
}
