% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_column.R
\name{test_column}
\alias{test_column}
\title{Runs Validation Checks on Columns names and number}
\usage{
test_column(df)
}
\arguments{
\item{df}{data frame to test}
}
\description{
Validate Scenario Modeling Hub submissions: names and number of columns.
}
\details{
This functions contains 2 tests:
\itemize{
 \item{"Name": }{The names of the columns are corresponding to the expected
 names: "model_projection_date", "scenario_name", "scenario_id", "target",
 "target_end_date", "location", "type", "quantile", "value". If one column
 is misspelled or is missing, the validation will stop here with an error
 message and no other tests will be perform on the submission.}
 \item{"Number": }{The submission should contains 9 columns.}
}
Function called in the `validate_submission()` function.
}
