% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_origindate.R
\name{test_origindate}
\alias{test_origindate}
\title{Runs Validation Checks on the Origin Date column}
\usage{
test_origindate(df, path, id)
}
\arguments{
\item{df}{data frame to test}

\item{path}{character vector path of the file being tested}

\item{id}{character date "id" of the corresponding round}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
\code{origin} column contains one unique date value corresponding
to the expected value, in the expected format.
}
\details{
This function contains 3 tests:
\itemize{
\item{Unique value: }{The \code{origin_date} column contains an unique expected
value.}
\item{Format: }{The \code{origin_date}
column contains a unique date value in either a character ("YYYY-MM-DD") or
a Date format. If the date is in a datetime format, a warning will be
returned}
}

Function called in the \code{validate_submission()} function.
}
