% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_origindate.R
\name{test_origindate}
\alias{test_origindate}
\title{Runs Validation Checks on the Origin Date column}
\usage{
test_origindate(df, path, id)
}
\arguments{
\item{df}{data frame to test}

\item{path}{character vector path of the file being tested}

\item{id}{character date "id" of the corresponding round}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`origin` column contains one unique date value corresponding
to the projection start date of the round and also corresponds to the date
in the name of the submission file.
}
\details{
This function contains 4 tests:
\itemize{
 \item{Date value: }{The `origin_date` column contains date value}
 \item{Unique value: }{The `origin_date` column contains a unique
 value.}
 \item{Correspondance to the name of the file: }{The `origin_date`
 column contains a unique date value matching the date in the name of the
 submission file}
  \item{Correspondance to the projection starting date: }{The
 `origin_date` column contains a unique date value matching the
 projection starting date of the corresponding round}
}
Function called in the `validate_submission()` function.
}
