% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_quantile.R
\name{test_quantiles}
\alias{test_quantiles}
\title{Runs Validation Checks on the Quantiles type and value columns}
\usage{
test_quantiles(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{js_def}{list containing round definitions: number and names of columns,
target names, ...}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`quantile` and `type` columns contain the expected information and value
and the projection value increases with the quantiles.
}
\details{
This function contains 4 tests:
\itemize{
 \item{Quantiles: }{The `quantile` column matches the expected quantiles}
 \item{Required quantiles: }{The projection should contain all required
  quantiles. It is accepted to submit less quantiles but the
  function will return a warning and the submission will not be included
  in the Ensembles.}
 \item{Value: }{The `value` associated with each `quantile` is increasing as
 the quantile increased. For example, if quantile 0.01 = 5 than quantile
 0.5 should be equal or greater than 5.}
 \item{Target: }{All targets required to have quantiles information have all
 required quantiles for each group id (scenario/location/horizon/etc.)}
}
Function called in the `validate_submission()` function.
}
