% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_req_value.R
\name{test_req_value}
\alias{test_req_value}
\title{Runs Validation Checks on the required value}
\usage{
test_req_value(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)

\code{model_task} should match a specific round model tasks from the
\code{tasks.json} associated with the hub. The json is expected to follow
the \href{https://hubverse.io/en/latest/user-guide/hub-config.html}{hubverse}
schema at least version 5.0

As the function was deprecated, it will not be updated anymore. It was
updated a last time to match the 5.0 hubverse schema version.}
}
\description{
\strong{DEPRECATED} \if{html}{\out{<br>}}\if{html}{\out{<br>}}
Validate Scenario Modeling Hub submissions: test if all the required value
are present in the submission file
}
