% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sample.R
\name{test_sample}
\alias{test_sample}
\title{Runs Validation Checks on the samplecolumn}
\usage{
test_sample(df)
}
\arguments{
\item{df}{data frame to test}
}
\description{
Validate Scenario Modeling Hub submissions: test if the  `sample` column
contains the expected  value.
}
\details{
This function contains 1 test:
\itemize{
 \item{sample value: }{The submission should contain a sample column with
 value between 0 and 100.}
}
Function called in the `validate_submission()` function.
}
