% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scenario.R
\name{test_scenario}
\alias{test_scenario}
\title{Runs Validation Checks on Scenario ID and Scenario Name columns}
\usage{
test_scenario(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
Validate Scenario Modeling Hub submissions: test if the name and id of the
scenario correspond to the expected name and id and if they are correctly
associated with each others.
}
\details{
This function contains 2 tests:
\itemize{
 \item{Scenario ID: }{The IDs of the scenarios are correctly spelled and
 correspond to the scenario IDs of the corresponding round.}
 \item{Required scenario: }{All the required scenario are present in the
 submission file (tested by target).}
}
Function called in the `validate_submission()` function.
}
