% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_target.R
\name{test_target}
\alias{test_target}
\title{Runs Validation Checks on the Target column}
\usage{
test_target(df, model_task)
}
\arguments{
\item{df}{data frame to test}

\item{model_task}{list containing round information for each id columns
and model output (type, format, etc.)}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`target` column contain the expected information and value.
}
\details{
This function contains 5 tests:
\itemize{
 \item{Target name: }{The target should correspond to the target name as
 expressed in the SMH Github. }
 \item{Target number: }{The submission file contains projection for all
 required targets. It is accepted to submit only a subset of target but
 a warning message will be return (only if a required target is missing.}
 \item{Number of week projected: }{The submission file contains projections
 for the expected number of weeks or more. If a team submits more week than
 expected, a warning message will be returned, but the submission will be
 accepted. However, an error message will be returned if the submission file
 contains less projection week than expected. Starting round 13, if the file
  contains less projected weeks than expected, the submission will still be
  accepted, but will return a warning message and might not be included
  in the Ensembles}
 \item{Week projected: }{The submission file contains projection for all
 expected weeks for each target, location, scenario (age_group) and quantiles
 or sample combination.}
 \item{NA target}{The projection contains NA for all "horizon"
 for the target requiring no time series information. }
}
Function called in the `validate_submission()` function.
}
