% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_target.R
\name{test_target}
\alias{test_target}
\title{Runs Validation Checks on the Target and target_end_date columns}
\usage{
test_target(df, start_date, round)
}
\arguments{
\item{df}{data frame to test}

\item{start_date}{corresponds to the "1 wk ahead" target in the projection
file}

\item{round}{numeric corresponding to the current round number}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`target` and `target_end_date` columns contain the expected information
and value.
}
\details{
This function contains 7 tests:
\itemize{
 \item{Target name: }{The target should correspond to the target name as
 expressed in the SMH Github: "inc death", "inc case", "cum death",
 "cum case", "inc hosp", "cum hosp". Starting round 14, an optional target
 is also possible "inc inf". }
 \item{Target number: }{The submission file contains projection for all
 required targets. It is accepted to submit only a subset of target but
 a warning message will be return (only if a required target is missing.}
 \item{End of week: }{The date in the "target_end_date" column should always
 correspond to the end of the epiweek (Saturday).}
 \item{Number of week projected: }{The submission file contains projections
 for the expected number of weeks or more. If a team submits more week than
 expected, a warning message will be returned, but the submission will be
 accepted. However, an error message will be returned if the submission file
 contains less projection week than expected. Starting round 13, if the file
  contains less projected weeks than expected, the submission will still be
  accepted, but will return a warning message and might not be included
  in the Ensembles}
 \item{Week projected: }{The submission file contains projection for all
 expected weeks for each target, location, scenario and quantiles
 combination.}
 \item{Starting date: }{The 1 week ahead of the `target_end_date` starts on
 the expected date (end of the epiweek of the starting projection date).}
 \item{Correct date: }{Each target_end_date corresponds to the expected date
  for example if 1 wk ahead = 2022-01-15, than 2 wk ahead = 2022-01-22.}
 \item{Target value: }{For the round 14 and the optional target "prop_x", the
  associated value with this target should be between 0 and 1.}
}
Function called in the `validate_submission()` function.
}
