% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_val.R
\name{test_val}
\alias{test_val}
\title{Runs Validation Checks on the Projection value and type point columns}
\usage{
test_val(df, pop, last_lst_gs, number2location)
}
\arguments{
\item{df}{data frame to test}

\item{pop}{data frame containing the population size of each geographical
entities by fips (in a column "location")}

\item{last_lst_gs}{list of data frame, named with the corresponding target and
containing the last avaible week of observed data  before start of the
projection}

\item{number2location}{named vector containing the FIPS as name and the
corresponding location name as value (example: name: "01", value: "Alabama")}
}
\description{
Validate Scenario Modeling Hub submissions: test if the
`value` and `type` columns contain the expected information and value.
}
\details{
This function contains 10 tests:
\itemize{
 \item{Type: }{The projection contains "point" value noted as `type` =
 "point".}
 \item{Quantile: }{The projection contains "point" value noted as `quantile`
 = NA.}
 \item{Number of Point: }{The projection contains X point value, X = number
 of targets * number of scenarios * number of locations}
 \item{Positive Value: }{The projection contains only values superior or
 equal to 0}
 \item{Unique Point: }{In the projection, each group of
 scenario/location/target has 1 unique point value (for the required targets)}
 \item{Unique value: }{For each target/scenario/location group (except
 locations 66 (Guam), 69 (Northern Mariana Island), 60 (American Samoa),
 74 (US. Minor Outlying Islands)), the whole time frame projection does not
 contain only 1 unique value. For example, contains only 0 or same
 cumulative value for the whole projections. The submission is still
 accepted if so, but will return a warning asking to verify if the projection
  is correct for the specific location/target group}
 \item{Value < Population size: }{Each projected value cannot by higher than
 the population size of the corresponding geographical entity (currently
 developed on each incidence and cumulative value, starting round 13). The
 submission is still accepted if the test failed, but will return a message
 asking to verify the projected value for the specific target and location}
 \item{Cumulative cases: }{In the submission, the projected cumulative
 case counts value should not be lower than the week 0 (or week -1, depending
  on availability on the time of submission) of the observed cumulative
  cases}
 \item{Cumulative deaths: }{In the submission, the projected cumulative
 death counts value should not be lower than the week 0 (or week -1,
 depending on availability on the time of submission) of the observed
 cumulative deaths}
 \item{Unique Projections: }{In the submission, each
 quantile/target/scenario/location combination has one unique value
 projected. (For example: only 1 value for quantile 0.5, location US,
 target 1 wk ahead inc case and scenario A).}
}
Function called in the `validate_submission()` function.
}
