% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_part_file.R
\name{validate_part_file}
\alias{validate_part_file}
\title{Validate Partitioned files level properties}
\usage{
validate_part_file(hub_path, folder_path, partition)
}
\arguments{
\item{hub_path}{path to the repository contains the submissions files
and \code{tasks.json} file.}

\item{folder_path}{path to the folder containing partitioned submission
files. The path should be relative to the \code{hub_path}, model output folder.}

\item{partition}{character vector corresponding to the column names of each
path segments}
}
\description{
Function based on \code{hubValidations::validate_model_file()}, adapted to work
for partitioned SMH file format
}
\details{
See \code{hubValidations::validate_model_file()} on the tests performed in the
function. Addtional tests includes checking the partition names and value
are matching the expected values.
}
\examples{
hub_path <- system.file("extdata/exp/", package = "SMHvalidation")
path_2 <- "t3-mc"
SMHvalidation::validate_part_file(hub_path, path_2,
                                  c("origin_date", "target"))

}
