% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_submission.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate SMH (Scenario Modeling Hub) Submissions}
\usage{
validate_submission(
  path,
  js_def,
  lst_gs,
  pop_path,
  merge_sample_col = FALSE,
  partition = NULL,
  n_decimal = NULL,
  round_id = NULL
)
}
\arguments{
\item{path}{path to the submissions file to test or string of parquet files (
in this case, the validation will be run on the aggregation of all the
parquet files together)}

\item{js_def}{path to JSON file containing round definitions: names of
columns, target names, ...}

\item{lst_gs}{named list of data frame containing the
observed data. For COVID-19, we highly recommend to use the output of the
pull_gs_data() function. The list should have the same format: each data
frame should be named with the corresponding covidcast signal except
"hospitalization" instead of "confirmed_admissions_covid_1d".}

\item{pop_path}{path to a table containing the population size of each
geographical entities by FIPS (in a column "location") and by location name.}

\item{merge_sample_col}{boolean to indicate if the for the output type
"sample", the output_type_id column is set to NA and the information is
contained into 2 columns: "run_grouping" and "stochastic_run". By default,
`FALSE`}

\item{partition}{vector, for csv and parquet files, allow to validate files
in a partition format, see `arrow` package for more information, and
`arrow::write_dataset()`, `arrow::open_dataset()` functions.}

\item{n_decimal}{integer,  number of decimal point accepted in the column
value (only for "sample" output type), if NULL (default) no limit expected.}

\item{round_id}{character string, round identifier. If `NULL` (default),
extracted from `path`.}
}
\description{
Runs all the different validation checks functions (test_column,
test_scenario, test_modelprojdate, test_quantiles, test_val, test_target,
test_location, test_sample, test_agegroup) on a Scenario Modeling Hub (SMH)
submissions and prints information about the results of each tests on the
submission: warning(s), error(s) or message if all the tests were successful.
}
\details{
For more information on all tests run on the submission, please refer
to the documentation of each "test_*" function. A vignette with all the
information is available in the package and is called: "validation-checks".

The function accepts submission in PARQUET, CSV, ZIP or GZ file formats.

The function runs some preliminary tests before calling the "test_*"
functions:
\itemize{
 \item{Input submission file format: }{The file format of the submission
 file(s) correspond to the expected format (for example: `parquet`, or `csv`,
  etc.). If multiple files inputted, only `parquet` is accepted}
  \item{Date information: }{The column `origin_date` in the submission file
  corresponds to a `model_tasks` round information in the JSON file
  (`js_def` parameter)}
  \item{Date format: }{All columns containing dates information should be in
  "YYY-MM-DD" format}
}
}
\examples{
\dontrun{
# FOR SMH Submission

lst_gs <- pull_gs_data()
pop_path <- "PATH/TO/data-locations/locations.csv"
js_def <- "PATH/TO/tasks.json"

validate_submission("PATH/SUBMISSION", js_def, lst_gs, pop_path)

}
}
