% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_submission.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate SMH (Scenario Modeling Hub) Submissions}
\usage{
validate_submission(
  path,
  hub_path,
  js_def = NULL,
  target_data = NULL,
  pop_path = NULL,
  merge_sample_col = NULL,
  partition = NULL,
  n_decimal = NULL,
  round_id = NULL,
  verbose = TRUE,
  r_schema = NULL
)
}
\arguments{
\item{path}{path to the submissions file (or folder for partitioned data)
to test, or string of parquet files (in this case, the validation will be
run on the aggregation of all the parquet files together, and each file
individually should match the expected SMH standard). The path should be
relative to the \code{hub_path}
If partition is not set to \code{NULL}, path to the folder containing ONLY the
partitioned data.}

\item{hub_path}{path to the repository contains the submissions files
and \code{tasks.json} file}

\item{js_def}{path to a JSON file containing round definitions: names of
columns, target names, ... following the \code{tasks.json}
\href{https://hubverse.io/en/latest/user-guide/hub-config.html}{Hubverse}
format. If \code{NULL} (default), will use the default path
"hub-config/tasks.json" in the hub using \code{hub_path} parameter.}

\item{target_data}{path to the target data in the Hubverse time series target
data standard format. Please find additional information on the
\href{https://hubverse.io/en/latest/user-guide/target-data.html}{hubverse}
website. Set to \code{NULL} (default), to NOT include comparison with observed
data.}

\item{pop_path}{path to a table containing the population size of each
geographical entities by FIPS (in a column \code{"location"}).
Use to compare that value is not higher than expected population size.
Set to \code{NULL} (default), to NOT run comparison with population data.}

\item{merge_sample_col}{vector to indicate if the for the output type
"sample", the output_type_id column is set to NA and the information is
contained into other columns. If so the parameter should be set to the sample
ID columns names, for example: \verb{c("run_grouping" and "stochastic_run")}.
By default, \code{NULL} (sample ID information in the output type id column)}

\item{partition}{vector, for csv and parquet files, allow to validate files
in a partition format, see \code{arrow} package for more information, and
\code{arrow::write_dataset()}, \code{arrow::open_dataset()} functions.  By default
\code{NULL}, no partition}

\item{n_decimal}{integer, number of decimal point accepted in the column
value (only for \code{"sample"} output type), if \code{NULL} (default) no limit
expected.}

\item{round_id}{character string, round identifier. If \code{NULL} (default),
extracted from \code{path}.}

\item{verbose}{Boolean, if TRUE add information about the sample pairing
information in output message. By default, \code{TRUE} (slows the validation
validation for sample output type)}

\item{r_schema}{Schema arrow objects, to read partition files with a specific
schema. If none provided (default, \code{NULL}), the schema will be created from
the \code{js_def} JSON file. (only for partition files)}
}
\description{
Runs multiple validation checks functions on SMH submission
file(s) and prints information about the results of each tests on the
submission file(s): warning(s), error(s) and/or message in the hubverse
format.
}
\details{
For more information on all tests run on the submission, please refer
to the documentation of the
\href{https://hubverse-org.github.io/hubValidations/index.html}{hubValidations}
package. A vignette with all the information is available in the package and
is called: vignette("validation-checks").

The function accepts submission in PARQUET, CSV, ZIP or GZ file formats. If
the submission files is in a "partitioned" format, the \code{path} parameter
should be to a directory to a folder containing ONLY the "partitioned"
files. If any other file is present in the directory, it will be included
in the validation.

The function runs some preliminary tests before running all the checks:
\itemize{
\item Input submission file format: The file format of the submission
file(s) correspond to the expected format (for example: \code{parquet}, or \code{csv},
etc.). If multiple files inputted, only \code{parquet} is accepted.
\item Date information: The column \code{origin_date} in the submission file
corresponds to a \code{model_tasks} round information in the JSON file
(\code{js_def} parameter).
\item Date format: All columns containing dates information should be in
"YYY-MM-DD" format.
\item Location name: The submission should contains projection by
location, the \code{location} column contains the location FIPS number as
available in the location table in the SMH GitHub Repository. If the FIPS
number are missing a trailing zero, the submission will be accepted but a
warning message will be returned.
\item Column format: If the submission file(s) contain column in a factor column,
the column will be forced to character and a warning message will be
returned
}
}
\examples{
\dontrun{
validate_submission("PATH/TO/SUBMISSION", "PATH/TO/HUB")
}
}
