% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_submission.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate SMH (Scenario Modeling Hub) Submissions}
\usage{
validate_submission(path, lst_gs, pop_path, js_def)
}
\arguments{
\item{path}{path to the submissions file to test}

\item{lst_gs}{named list of data frame containing the
observed data. For COVID-19, we highly recommend to use the output of the
pull_gs_data() function. The list should have the same format: each data
frame should be named with the corresponding covidcast signal except
"hospitalization" instead of "confirmed_admissions_covid_1d".}

\item{pop_path}{path to a table containing the population size of each
geographical entities by FIPS (in a column "location") and by location name.}

\item{js_def}{list containing round definitions: names of columns,
target names, ...}
}
\description{
Runs all the different validation checks functions (test_column,
test_scenario, test_modelprojdate, test_quantiles, test_val, test_target,
test_location, test_sample, test_agegroup) on a Scenario Modeling Hub (SMH)
submissions and prints information about the results of each tests on the
submission: warning(s), error(s) or message if all the tests were successful.
}
\details{
For more information on all tests run on the submission, please refer
to the documentation of each "test_*" function. A vignette with all the
information might be created later on too.
\cr\cr
The function accepts submission in CSV, ZIP or GZ file formats.
}
\examples{
\dontrun{
# FOR COVID-19 SMH Submission

lst_gs <- pull_gs_data()
pop_path <- "PATH/TO/data-locations/locations.csv"
js_def <- jsonlite::fromJSON("PATH/TO/covid.json")

validate_submission("PATH/SUBMISSION", lst_gs, pop_path, js_def)

# FOR FLU SMH Submission

lst_gs <- NULL
pop_path <- "PATH/TO/data-locations/locations.csv"
js_def <- jsonlite::fromJSON("PATH/TO/flu.json")

validate_submission("PATH/SUBMISSION", lst_gs, pop_path, js_def)

}
}
